feat(#38): 힌트 시간 상한선 - 힌트 시간 상한선 설정 100초로 제한 - sumPlayTimeByMapId 반환 타입 수정

+++ src/main/java/io/github/ascrew/monomatbe/domain/map/repository/MapItemJpaRepository.java
@@ -25,5 +25,5 @@ public interface MapItemJpaRepository extends JpaRepository<MapItem, Long> {
             from MapItem mi
             where mi.map.id = :mapId and mi.isDeleted = false
             """)
-    Integer sumPlayTimeByMapId(@Param("mapId") Long mapId);
+    Long sumPlayTimeByMapId(@Param("mapId") Long mapId);
 }
